refactor: 파일명 컨벤션 적용 (#9)
