+++ src/components/draw-banner.tsx
