+++ src/components/guide-section.tsx
