+++ src/components/sparkle-burst.tsx
