+++ src/data/ornament-message.ts
